langchain/base.py
from dotenv import load_dotenv
load_dotenv()

from langchain.chat_models import init_chat_model

model = init_chat_model("gpt-4o-mini", model_provider="openai")
output = model.invoke("自己紹介してください")
print(output.content)