requirements.txt
python-dotenv
openai
requests
langchain
langchain_openai
langchain_community
langchain_chroma 